requirements.txt
transformers
streamlit
python-dotenv
langchain
langchain-community
torch
sentence-transformers
bitsandbytes
accelerate
faiss-cpu
poetry
ipykernel